requirements.txt
opencv-python
numpy
matplotlib
geopandas
fiona
requests
python-dotenv
Pillow
fastapi
uvicorn[standard]
earthengine-api
pydantic
python-multipart
